fix(ptringboot-gestion-encuestas): se agrega la implementación de los servicios de gestión de encuestas.

+++ src/main/java/com/springbootgestionencuestas/springboot_gestion_encuestas/repository/EncuentaRepository.java
@@ -3,5 +3,5 @@ package com.springbootgestionencuestas.springboot_gestion_encuestas.repository;
 import com.springbootgestionencuestas.springboot_gestion_encuestas.model.Encuesta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EncuentaRepository extends JpaRepository<Encuesta, Integer> {
+public interface EncuentaRepository extends JpaRepository<Encuesta, Long> {
 }
