+++ src/main/java/com/springbootgestionencuestas/springboot_gestion_encuestas/repository/RespuestaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RespuestaRepository extends JpaRepository<Respuesta, Integer> {
+public interface RespuestaRepository extends JpaRepository<Respuesta, Long> {
     List<Respuesta> findByPregunta(Integer preguntaId);
 }
